requirements.txt
autopep8==1.6.0
cycler==0.11.0
flake8==4.0.1
fonttools==4.34.2
kiwisolver==1.4.3
matplotlib==3.5.2
mccabe==0.6.1
numpy==1.23.0
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
toml==0.10.2
